+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New feature: Added scene view overlay tools for Cinemachine components.
 - Regression fix: Lookahead works again.
 - Cinemachine3rdPersonAim exposes AimTarget, which is the position of where the player would hit.
+- Bugfix: Negative Near Clip Plane value is kept when camera is orthographic.
 
 
 ## [2.8.0] - 2021-07-13